Refactored the RPC call function into a separate file for easy reuse
